lots of bug-fixes that had problems with pbc migration
The boundary conditions were rarely properly
shipped together with the Lattice methods.
Basically it boiled down to the copy method
which did not take all variables into account.

This commit also addes a Listify|listify
class|object that can be used to ensures
arguments are converted to a list.
It can be quite handy as it allows "piping".

Enabled a setter for pbc to fast changing
stuff. One can pass *none* to not change it.

Signed-off-by: Nick Papior <nickpapior@gmail.com>