fixed hsxSileSiesta reading from fdf siles
The problem was that the geometrry was only
read from the binary file, this meant that
the basis could never contain the orbital
radial parts etc.
This fixes this issue by trying to read
the basis and pass this down the read
structure.

Ideally one should create merge strategies
for the orbitals/atoms etc. to make it more
versatile and simpler to use, that has
to wait.

Also fixed quite a bit of typing.

Removed Orbital.radii (equivalent to radius).

Signed-off-by: Nick Papior <nickpapior@gmail.com>