completed tests rearrangement
Now all tests are following the structure
proposed in sisl-files #14.

It means complete output storage, and
cherry-picks for the files to be tested.

This of course meant a huge restructuring of tests
since tests had to be redone, re-arranged etc.

In particular, we have changed the tbtrans test
to be much smaller, and hence that particular
test is much faster!

Signed-off-by: Nick Papior <nickpapior@gmail.com>