fixed eigenmode reader, state order changed
Also ensured that atol is used (not ktol).

read_eigen* methods which can't find stuff based
specific k-values now raises LookupError,
not a KeyError, not an IndexError.

Cleaned tests for the vibra method.
Overall cleaning of the class as well.
Added a read_brillouinzone in the vibra output.

Removed sc argument allowance in eigenmode.
This is deprecated for other methods as well.

Signed-off-by: Nick Papior <nickpapior@gmail.com>