fixed reading geometries from fdf + XV + STRUCT
The different geometries reading mechanisms
where kind-of broken for various constalletations
in the XV+fdf+struct files.

The species_as_Z was basicaly never going to
works since there was no mechanism to counter
it.

Instead the XV and STRUCT siles now handle
basis from the perspective of being passed
an atoms argument, from which it will
fill in stuff based on data.
If the atoms is a *pure* basis object (only
correct species indices).
Then it will replace without checking stuff.
If it has the same length as the geometry
it will check Z and whether indices are
the same. This *can* potentially lead
to errors if there is only 1 atom of each
species, but on the other hand the species
order should then handle the indices correctly.

There was also cases in fdf, XV and STRUCT reads
that could potentially lead to reducing the basis.
We should generally try to avoid this since
there may be good reasons for having *empty* basis.

Signed-off-by: Nick Papior <nickpapior@gmail.com>