fixed lots of parallel calculations
A lot of refactoring of the parallel codes
enables huge speedups.

The main culprit for bad scaling is the chunksize.

A new environment variable has been introduced:

   SISL_PAR_CHUNKSIZE=25 | 0.1

it specifies the default chunksize for the pool.*()
methods.
It defaults to a fractional number in which case it means the number
of inverse chunks each CPU gets, i.e. 0.1 == 10 chunks per processor.
Generally this shows a good scaling while fine-tuning
can generally help.

I can now see huge parallel performance benefits by leveraging
the chunksize variable.

The dispatcher for the `bz.apply` method now allows a finer
tuning of the pool creation.

1. Single argument

    bz.apply(pool=2)

   Will create a pool with 2 processors

2. Two tuple

    bz.apply(pool=(2, {"chunksize": 200}))

   will create a pool of 2 processors, and the chunksize
   will be 200 (regardless of SISL_PAR_CHUNKSIZE)

3. Three tuple

    bz.apply(pool=(2, {"args": 1}, {"chunksize": 200}))

   will create a pool of 2 processors like so:

       pathos.pools.ProcessPoll(nodes=2, args=1)

   check the documentation for `pathos` to see what can be done.
   Generally this need not be used.

These 3 variants are currently added, but I would like some input
to see whether it makes sense, or whether we should change arguments etc.

The default number of processers is still 1, this is because the OMP_NUM_THREADS
can easily create deadlocks if SISL_NUM_PROCS * OMP_NUM_THREADS > CORES.
So to be on the safe side, we default it to 1.

Since parallel processing is now by default *on*, one should simply
need to do:

   SISL_NUM_PROCS=2 SISL_PAR_CHUNKSIZE=50 python3 script.py

then the procs are determined from the variable whenever a bz.apply is found.
In addition, all parallel invocations will now correctly
update the progressbars from tqdm.

Since many routines might loop over distribution functions, we have
changed them to enable broad-casting (internally).
This required us to bump the required numpy version to >=1.20.
This release is from January 2021.

Many dot calls has been changed to @, numpy recommends matmul when
that is the intention. The primary reason is that scipy.sparse.dot(np.ndarray)
can in certain cases result in an np.ndarray with dtype=object where the
actual elements are actually a sparsematrix. So we can't use that.

There are still some corner cases where @ cannot be used.
E.g. 1 @ array will fail, it does not work on scalars. This is a bit
unfortunate as it would ease things a bit.

Added more typing in state.py, electron.py and some minor other places.

Signed-off-by: Nick Papior <nickpapior@gmail.com>