Clarify out of memory error on density
I have been testing the limits and for 10000 orbitals numpy needs to allocate 23GiB, which is too much for my RAM, and it fails.

It then throws a `numpy.core._exceptions._ArrayMemoryError`, which a user might not understand how to solve. Here we raise a clearer error for the user that is trying to compute the density.

It takes veeery long to compute such a system with the direct method, but at least it can be computed.