ahc and shc calculations enabled
after #766 these changes were made.

- `berry_curvature` is changed to only live on
  the `StateCElectron`.
  Old scripts relying on `berry_curvature`
  will not work now.
- `conductivity` has been deprecated, in favor
  of explicit calculations, `ahc` and `shc`.
- `axes` and `operator` arguments has been added
  for `derivative`
  They allow to determine which axes to extract the derivatives
  of, and to change the velocity operator to something else.

The above changes are required to enable the `ahc` and `shc` calculations.

ahc is basically the same as the old `conductivity` with changed API.
shc is the ahc where one of the velocity operators is changed
to the spin current operator.

Lots of the methods has changed their API's to enable faster parallel
calculations by explicitly allowing arguments passed to apply
methods.

However, the units of these methods are still wrong.
That needs subsequent fixing.

Signed-off-by: Nick Papior <nickpapior@gmail.com>