enabled k-resolved ahc and shc calculations
These will then be very memory requiring.

Changed algorithm to return an imaginary number.
The real quantity should be the ahc/shc. But
it might be needed for other stuff.
E.g. when we add the alternate current stuff
with hbar omega terms.

Reversed the distribution functions which added
a factor -1.

Signed-off-by: Nick Papior <nickpapior@gmail.com>