enabled warnings for EF when not available for HSX files
This caused a confusion since it was only fdf siles
that raised this problem. This is now fixed because
hsx files will *always* raise, then the caller should catch
them. Probably other places have this issue as well.

We might even consider doing this for basis sets etc.

Also, fixed a minor error when tracking fdf and
reading Hamiltonian from TSHS (an error in argument).

Signed-off-by: Nick Papior <nickpapior@gmail.com>