fixed Eindex usage, int same behavior as float
The tbt.Eindex method has been confusing for too
long.
Now Eindex is working as it *should*.
*Any* value used, will be regarded as a float.
It streamlines the usage, and also allows users to
do `Eindex(0)` to get the energy index corresponding
to the energy closests to the reference energy.

This will definitely break somebodys workflows, but
likely for the better.

The `Eindex` also got a new argument, `method`
which can accept nearest, below, above which determines
the direction of energy grabbing. nearest is default.

A lot of typing has also been added to the tbtrans
methods.
It should make the documentation a bit clearer.

Signed-off-by: Nick Papior <nickpapior@gmail.com>