first attempt at fixing the documentation builds
This is just the first of some more fixes that should
improve the documentation of the dispatchers used
in sisl.

Basically things are hard to change in sphinx. The simplest
way is really to do it in the jinja templates.
The templates has some hard-coded details in them.
Which is kind of bad... On the other hand, it works nicely
because all the dispatchers has unique names.
Only when new dispatchers are used, should we amend specific
sections, or join them elsewhere (if needed).

So, the class template is very verbose with lots of
macros, and actual code. It is not ideal, on the other
hand it is self-contained and relatively simple to
understand (except the call(...) macro_call()).

The template changes was also the simplest way to
manually control the order of segments.
Probably this requires some finetuning.

Signed-off-by: Nick Papior <nickpapior@gmail.com>