fixing sphinx builds by properly hidding details
The problem is that the :hidden: qualifier is not
allowed for autosummary directives.

Instead, we now have the autosummary (without hidden)
in separate files, suffixed with -hidden.rst to hide them.
To disable warnings here they are marked with :orphan:.

Lastly, a bug appeared in the v0.15.0 release, this is now also fixed.

Signed-off-by: Nick Papior <nickpapior@gmail.com>