Small clean-ups in conf.py
This mainly relates to allowing to build
faster (no ipynb builds

Do
_SISL_DOC_SKIP=notebook sphinx ...

to skip notebook stuff, should be amended
for further reducing build times (can
be very handy when just checking single
pages).

Added a comment for the duration extension.
This will show which classes takes the longests
to build.

Signed-off-by: Nick Papior <nickpapior@gmail.com>