Accessors and grouping documentation entries (#845)
* first attempt at fixing the documentation builds

This is just the first of some more fixes that should
improve the documentation of the dispatchers used
in sisl.

Basically things are hard to change in sphinx. The simplest
way is really to do it in the jinja templates.
The templates has some hard-coded details in them.
Which is kind of bad... On the other hand, it works nicely
because all the dispatchers has unique names.
Only when new dispatchers are used, should we amend specific
sections, or join them elsewhere (if needed).

So, the class template is very verbose with lots of
macros, and actual code. It is not ideal, on the other
hand it is self-contained and relatively simple to
understand (except the call(...) macro_call()).

The template changes was also the simplest way to
manually control the order of segments.
Probably this requires some finetuning.

I still can't do it fully programmatically, that would
require a very dense jinja template.
I think this is *ok* for now. Perhaps some discovery of
which flags are really used is necessary.

* fixed SileBound documentation

It encompasses nicely with the way this accessor thing
patches the documentation.

* safety net in doc creation when adding more dispatchers

This should warn when adding new dispatchers, and also
indicate exactly which one is missing.

Also made nice use of the whalrus op.

* added decorator to streamline process of creating classes all the time

* use inspect.getdoc instead of manual

* implemented @pfebrer's suggestion for customizing names

This is simple, and works.

Signed-off-by: Nick Papior <nickpapior@gmail.com>

---------

Signed-off-by: Nick Papior <nickpapior@gmail.com>