should fix the reading of the stresses and forces
Done to fix #832.

This should consistently read the last stress/force.

Now stdoutSileSiesta.read_stress|force
behaves consistently:

If they are found in a dynamics section, i.e.
before the Final section, then the stress/forces
in the Final section are discarded as entries.

If they are not found in the dynamics section,
then the things in the Final section will be returned.

In a user explicitly only wants the final section,

do read_stress(skip_final=False)

This required some restructure in the InfoAttr handling.
But also found some weaknesses there. I think it works
a bit better now.

Lastly, added lots of typing.

Signed-off-by: Nick Papior <nickpapior@gmail.com>