fixed 5.2 reading of charges
Now the stdout sile for siesta tries
to automatically determine the used charge
reading algorithm, one can still forcefully
request the old-style. But if the version
is 5.2 or later, it will default
to the new scheme.

This commit also fixes the issue of the changed
column names.

The stdout sile now also contains the version
specification.

Fixes #856

Signed-off-by: Nick Papior <nickpapior@gmail.com>