initial commit of nambu code (#877)
* initial commit of nambu code

Added matrix codes to create the Nambu
Hamiltonian.
This required restructuring the diag
fold_csr matrix codes. Now they can be given
a number to determine the number of elements
that are added. It generalizes the code a bit.

Currently the matrix creation does not implement
the no-phase code-path. So Gamma-point is not functional.

Changed some of the cython codes to use preprocessors
at the comment level at the top of the file.
It makes it much simpler to debug.

Enabled siesta routines to read the matrices
with Nambu spin configuration.

Signed-off-by: Nick Papior <nickpapior@gmail.com>

* finalized the PDOS for nambu and finalized more methods

- nambu PDOS
- nambu transpose
- nambu berry-phase stuff works
- nambu trs is *NOT* implemented.

Added more tests for complex data-types which
has been completed in this branch.

* Added warning when using nambu spin + changelog

Signed-off-by: Nick Papior <nickpapior@gmail.com>

---------

Signed-off-by: Nick Papior <nickpapior@gmail.com>