enabled writing HSX files
This not only allows writing HSX files, but also
completes the version == 1, version == 2 specification.

The later version is going to be the new default
for transiesta runs, given that it contains all information
that the TSHS file contains + more.

Tests are added to ensure they are consistent.

However, the charges for the orbitals are not correctly
stored, that should be solved later on.

Signed-off-by: Nick Papior <nickpapior@gmail.com>