enabled ufuncs for external classes (#898)
* enabled ufuncs for external classes

This enables one to interact on external classes.
However, only if they can be converted to a sisl object.

This can be quite handy for sisl specific functions.

Fixes #747.

added a tip in the functional programming documentation

Allowed one to forcefully return the sisl object.

Will not act on built-ins, because it can lead to simple ambiguity.

Signed-off-by: Nick Papior <nickpapior@gmail.com>

---------

Signed-off-by: Nick Papior <nickpapior@gmail.com>