Changed eta behavior in WideBandSE
eta cannot be used anymore, only E argument.
And only the imaginary part will be used.

The returned wide-band matrix now has a factor
np.pi, to ensure the integration is *correct*
in terms of the Lorentzian.

Finally, the WideBandSE will now use the overlap
matrix if it exists for the instantiated object.
This will mean that the Green/self-energy has
the same structure as a non-orthogonal basis
set.

Lots of typing has been fixed for the self_energy
file.

Signed-off-by: Nick Papior <nickpapior@gmail.com>