allowed *any* dtype for spin_rotate, spin_align, mulliken
This fixes #911.

This now uses the shared method _get_spin (currently internal)
to unify the process. Instead of doing things manually in each
method, we can rely on a single code that performs all necessary
actions related to the spin quantity. It might even
be useful to construct sparse matrices in vector space.

Signed-off-by: Nick Papior <nickpapior@gmail.com>