updated to CODATA 2022
Now CODATA 2022 is the default, and the implementation
allows to more easily swap between different CODATA versions.

Currently 2010, 2014, 2018 and 2022 are existing and usable.

Together with this a full restructure of how the unit table
is constructed has taken place. It more easily adds new
units, and is a specific class (more or less a dict extension).

This fixes #912.

Signed-off-by: Nick Papior <nickpapior@gmail.com>