restructuring the read_scf for stdoutSileSiesta
This fixes #918 which cause erroneous reads because
some details are stored *after* the scf: key.
Essentially the spin moment is consistently written
after.

The current approach in read_scf, is to read data until
it hits the same data-key *again*, at which point it
serializes the data and stores it for further processing.

The logic has been streamlined and it should be simpler to follow
and edit.

Signed-off-by: Nick Papior <nickpapior@gmail.com>