fixed and enhanced basis-optimizer
Added needed dependency for pyyaml.
While strictly not necessary, I think this is
an ok thing. It is very much used elsewhere, and
we might require it for other things.

Made an alias for stoolbox basis optimize|optim.

Fixed a thing for return numpy arrays in periodictable.

The new CLI registry requires handling of union types, this
will fix it by converting unions to str, regardless of what
it actually is.

Added information when basis information is requested but
the entries won't be there due to too small numbers.
E.g. soft confinement with V0 = 0 will effectively not do
anything. Same for charge confinement for abs(Z) = 0.

Added soft-confinement. Made charge-confinement available for
all orbitals with 0 charge. Soft-confinement are added
to all orbitals, which likely is overkill. It should probably
only be done for narrow orbitals. Well...

When *guessing* the orbitals I found some errors, at least
for gold, it simply omitted the p orbitals, so now there
is a warning, and then it will continue silently.
It's better to rely on user-defined basis-information.

Enabled reading psml files from SIESTA_PS_PATH (not psf!).

Signed-off-by: Nick Papior <nickpapior@gmail.com>