changed the bloch code to use fused dtypes
This just reduces maintenance burden.
No perf tests has been done, yet.

Cython is really annoying to maintain.
When there are obscure errors, e.g.
I have to pre-create the mem-view before
passing it, which is really annoying and
does not make sense.
Therefore it might be useful to try and
change the code to use another language.