Changed to scipy.sparse.*_array
This will be the preferred way of interacting
with scipy sparse matrices.

However, they have the down-side of not allowing
1D slicing (in older versions). Only in
scipy commit: 66ec33370c889e9c9acba320354b624ee258ee22
did this enter (in June 2024).

So for now, it has to be limited only, but with the latest
scipy, it shouldn't be a problem.